feat(usuario): adicionar validação de senha ao deletar e atualizar campos
- Deletar usuário agora exige senha para validação.
- Atualização de usuário agora considera todos os campos.
- Ajustes em entity: tipos, limites de tamanho e formatação da data de nascimento.

+++ src/main/java/com/api/meal4you/service/UsuarioService.java
@@ -4,7 +4,6 @@ import com.api.meal4you.entity.Usuario;
 import com.api.meal4you.repository.UsuarioRepository;
 import org.springframework.stereotype.Service;
 
-
 @Service
 public class UsuarioService {
     private final UsuarioRepository repository;
@@ -13,28 +12,38 @@ public class UsuarioService {
         this.repository = repository;
     }
 
-    public void salvarUsuario(Usuario usuario){
+    public void salvarUsuario(Usuario usuario) {
         repository.saveAndFlush(usuario);
     }
 
-    public Usuario buscarUsuarioPorEmail(String email){
+    public Usuario buscarUsuarioPorEmail(String email) {
         return repository.findByEmail(email).orElseThrow(
-                () -> new RuntimeException("Email não encontrado")
-        );
+                () -> new RuntimeException("Email não encontrado"));
     }
 
-    public void deletarUsuarioPorEmail(String email){
+    public void deletarUsuarioPorEmail(String email, String senha) {
+        Usuario usuario = repository.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("Email não encontrado"));
+        if (!usuario.getSenha().equals(senha)) {
+            throw new RuntimeException("Senha incorreta");
+        }
         repository.deleteByEmail(email);
     }
-    public void atualizarUsuarioPorId(int id, Usuario usuario){
-        Usuario usuarioEntity = repository.findById(id).orElseThrow(() ->
-                new RuntimeException("Usuario não encontrado"));
+
+    public void atualizarUsuarioPorId(int id, Usuario usuario) {
+        Usuario usuarioEntity = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Usuario não encontrado"));
         Usuario usuarioAtualizado = Usuario.builder()
-                .email(usuario.getEmail() != null ? usuario.getEmail() :
-                        usuarioEntity.getEmail())
-                .nome(usuario.getNome() != null ? usuario.getNome() :
-                        usuarioEntity.getNome())
                 .id(usuarioEntity.getId())
+                .email(usuario.getEmail() != null ? usuario.getEmail() : usuarioEntity.getEmail())
+                .nome(usuario.getNome() != null ? usuario.getNome() : usuarioEntity.getNome())
+                .senha(usuario.getSenha() != null ? usuario.getSenha() : usuarioEntity.getSenha())
+                .localizacao(
+                        usuario.getLocalizacao() != null ? usuario.getLocalizacao() : usuarioEntity.getLocalizacao())
+                .data_nascimento(usuario.getData_nascimento() != null ? usuario.getData_nascimento()
+                        : usuarioEntity.getData_nascimento())
+                .tempo_disponivel(usuario.getTempo_disponivel() != null ? usuario.getTempo_disponivel()
+                        : usuarioEntity.getTempo_disponivel())
                 .build();
 
         repository.saveAndFlush(usuarioAtualizado);
